Rewrite with Go
Go is easier to write and maintain, builds faster and has finally
improvements on how modules are managed.

This is not just a rewrite with a different language.

With the help of Go concurrency features, this new version pushes
objects concurrently, resulting in faster uploads.

License has been changed from GPL-3.0-or-later to AGPL-3.0-or-later.

We also follow REUSE specifications for copyright and license notices.